backend/backend/requirements.txt
asgiref==3.2.10
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==2.4.0
psycopg2-binary==2.8.6
Pillow==8.3.1
drf-extra-fields==3.4.1
simplejson==3.18.4
gunicorn==20.0.4
reportlab==3.6.6